fix(prejoin): copy meeting info
