+++ react/features/invite/components/add-people-dialog/web/index.js
@@ -1,4 +1,3 @@
 // @flow
 
 export { default as AddPeopleDialog } from './AddPeopleDialog';
-export * from './utils';
